Added Solr timeout override ➕😎⏰
We're still seeing significant lag time between changes by staff to metadata and the appearance of those changes in the PUI. One possibility is to make this change, so we're giving it a try. Strictly speaking, TEST is find right now, but we want to keep things in alignment wherever possible across environments.